--- resources/ts/Router.tsx
@@ -1,57 +1,0 @@
-import React from "react";
-import { BrowserRouter, Route, Routes } from "react-router-dom";
-import Header from "./views/components/header/Header";
-import Login from "./views/pages/auth/Login";
-import Register from "./views/pages/auth/Register";
-import VerifyEmail from "./views/pages/auth/VerifyEmail";
-import VerifyEmailComplate from "./views/pages/auth/VerifyEmailComplate";
-import ForgotPassword from "./views/pages/auth/ForgotPassword";
-import ForgotPasswordCompalate from "./views/pages/auth/ForgotPasswordCompalate";
-import ResetPassword from "./views/pages/auth/ResetPassword";
-import ResetPasswordComplate from "./views/pages/auth/ResetPasswordCompalate";
-import UserAccountSetting from "./views/pages/settings/UserAccountSetting";
-import EmailSettings from "./views/pages/settings/EmailSettings";
-import PasswordSettings from "./views/pages/settings/PasswordSettings";
-import UserNameSettings from "./views/pages/settings/UserNameSettings";
-import ProfileSettings from "./views/pages/settings/ProfileSettings";
-import TopPage from "./views/pages/top/TopPage";
-import PostDesk from "./views/pages/desks/PostDesk";
-
-const Router = () => {
-  return (
-    <BrowserRouter>
-      <Header />
-      <Routes>
-        {/* トップページ */}
-        <Route path="/" element={<TopPage />} />
-        {/* 認証ページ */}
-        <Route path="/login" element={<Login />} />
-        <Route path="/register" element={<Register />} />
-        <Route path="/verify/email" element={<VerifyEmail />} />
-        <Route
-          path="/email/authorization/complete"
-          element={<VerifyEmailComplate />}
-        />
-        <Route path="/forgot/password" element={<ForgotPassword />} />
-        <Route
-          path="/forgot/password/compalate"
-          element={<ForgotPasswordCompalate />}
-        />
-        <Route path="/api/reset-password" element={<ResetPassword />} />
-        <Route
-          path="/reset-password/compalate"
-          element={<ResetPasswordComplate />}
-        />
-        <Route path="/post" element={<PostDesk />} />
-        {/* 設定ページ */}
-        <Route path="/settings/account" element={<UserAccountSetting />} />
-        <Route path="/settings/name" element={<UserNameSettings />} />
-        <Route path="/settings/email" element={<EmailSettings />} />
-        <Route path="/settings/password" element={<PasswordSettings />} />
-        <Route path="/settings/profile" element={<ProfileSettings />} />
-      </Routes>
-    </BrowserRouter>
-  );
-};
-
-export default Router;
