+++ resources/ts/route/Router.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import { BrowserRouter } from "react-router-dom";
+import Header from "../views/components/header/Header";
+import UserRoute from "./UserRoute";
+import AdminRoute from "./AdminRoute";
+
+const Router = () => {
+  return (
+    <BrowserRouter>
+      <Header />
+      <UserRoute />
+      <AdminRoute />
+    </BrowserRouter>
+  );
+};
+
+export default Router;
